\name{specif}
\alias{specif}
\alias{specif-package}

\title{~ Overview: ... ~}
\description{
  \tabular{ll}{
    Version: \tab \Sexpr[stage=render]{library(utils);packageVersion("specif")}\cr
    Date: \tab \Sexpr[stage=build]{format(Sys.time(),"\%Y-\%m-\%d")}\cr
    License: \tab \Sexpr[stage=build]{library(utils);packageDescription("specif")$License} \cr
  }
}

\details{
  The package \code{specif} ...
}

\section{Author}{
  Bernard Desgraupes \cr
  \email{bernard.desgraupes@u-paris10.fr}\cr
  University of Paris Ouest - Nanterre - Lab Modal'X \cr \cr

  Sylvain Loiseau \cr
  \email{sylvain.loiseau@univ-paris13.fr} \cr
  University Paris-13 - Lab LDI \cr \cr
}

\references{
P. Lafon \emph{Sur la variabilité de la fréquence des formes dans un corpus}, Mots, octobre 1980, N°1. pp. 127-165.
}

\seealso{\code{\link{specif_calc}}}

\keyword{package}

